Do not include numbers in auto-generated NAME value.
When generating the NAME value from the provided variables, if the
variables include a value to be applied to `replicas` entry, then
changing the replica count would change the `name` as well which
causes Kubernetes to launch a new deployment, rather than scale
the existing one. This resolves that by excluding any value with
numbers from the values used to generate the name.